requirements.txt
tensorflow>=2.10.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0.0
kaggle>=1.5.0